requirements_docker.txt
# ============================================================
# Docker 容器依赖（Linux 环境，已排除 Windows 专用包）
# ============================================================

# Web 框架
fastapi==0.128.7
uvicorn==0.40.0
python-multipart==0.0.22
starlette==0.52.1

# Pydantic
pydantic==2.12.5
pydantic-settings

# OCR（paddlepaddle 在 python:3.10-slim 上需要手动安装）
paddlepaddle==3.0.0
paddleocr==3.3.3

# 图像处理（headless 版不依赖 GUI 库，服务器首选）
opencv-python-headless
pillow
numpy==1.26.4
shapely==2.1.2

# AI / LLM
openai==1.109.1
zhipuai==2.1.5.20250825
zai-sdk==0.2.2
langchain-openai==0.1.0
langchain-core==0.1.53

# 文档处理
python-docx==1.2.0
python-pptx==1.0.2
openpyxl==3.1.5
PyPDF2==3.0.1
lxml
pandas

# 工具库
PyYAML==6.0.2
python-dotenv==1.2.1
requests==2.32.5
httpx==0.28.1
tqdm==4.67.1
colorlog==6.10.1
websockets
orjson

# 其他依赖
sqlalchemy
tiktoken
